+++ src/ipcutil.h
@@ -17,7 +17,7 @@ extern "C" {
 
   /* allocate size bytes in shared memory with the specified flags and key.
      returns the pointer to the base address and the shmid, id */
-  void* ipc_alloc (key_t key, int size, int flag, int* id);
+  void* ipc_alloc (key_t key, size_t size, int flag, int* id);
 
   /* operate on the specified semaphore */
   int ipc_semop (int semid, short num, short incr, short flag);
